concierge/requirements.txt
fastapi[standard]>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
pyjwt[crypto]>=2.9.0
python-dotenv>=1.0.0
mcp>=1.0.0
